New 2d Game Components
The ultimate goal of this library is to simplify the creation of 2d game
development, and the latest addition is the first step to making that
possible. The Collidable and Entity classes provide users with simple
bases for creating in-game entities. They take care of collisions,
repositioning, movement, etc. More will be added to these classes in the
future.